pyproject.toml
[project]
name = "papis-pet"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "chromadb>=1.5.0",
    "fastapi>=0.128.8",
    "google-generativeai>=0.8.6",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "sentence-transformers>=5.2.2",
    "uvicorn>=0.40.0",
]
